requirements.txt
amqp==2.5.2
asgiref==3.2.7
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
billiard==3.6.3.0
blessed==1.17.5
boto3==1.13.7
botocore==1.15.49
cached-property==1.5.1
celery==4.4.2
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
Django==3.0.6
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-storages==1.9.1
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.17.1
future==0.16.0
idna==2.7
importlib-metadata==1.7.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
kombu==4.6.8
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.1.2
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycparser==2.20
pydot==1.4.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
pytest-django==3.9.0
python-crontab==2.4.2
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.1
requests==2.20.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
semantic-version==2.5.0
sentry-sdk==0.16.1
six==1.11.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
termcolor==1.1.0
texttable==1.6.2
uritemplate==3.0.1
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.1.0
